+++ meridian/__init__.py
@@ -14,7 +14,7 @@
 
 """Meridian API."""
 
-__version__ = "0.15.0"
+__version__ = "0.16.0"
 
 
 from meridian import analysis